adds SYM and BIN format inputs to m17-demod
Tried to figure out the logical way to inject symbols/dibits
into the demodulator, and it turned out to be fairly difficult.

The easier way to handle the symbols/dibits was to run them
through a modulator to generate baseband (rrc), and simply
run the baseband through the standard demodulator.

It's somewhat more brute force, but ended up being very simple
and clean.