Process first stream frame differently
M17 processes the LSF sync word differently than the STREAM sync word. Since we don't have an LSF, I combined these two cases, erroneously. There's an
important distinction: before the first frame, we don't
have frame timing. After the first frame, we need to
restrict our search for the sync word to a narrow
window around where it's supposed to be.

In exchange for this, we get to paper over missed sync
words. We still need to find out why.